Move new parser into parser.py
disable type checks in parser.py
