+++ mypy.ini
@@ -39,5 +39,7 @@ ignore_errors = True
 [mypy-pyquil.magic]
 ignore_errors = True
 
-[mypy-pyquil.parser2]
+# Ignore errors in the pyquil/parser.py file
+# TODO(notmgsk): Add types
+[mypy-pyquil.parser]
 ignore_errors = True
